Reorganization of the FDK
This patch aims to simplify the FDK by making a clear separation between the
files that the makers will have to handle and the FDK infrastructure.

The input files (namely the board support C files, config and manifest) that a
Maker should provide for a certain module are now contained in separate
directory which name is given as a parameter to the Makefile (e.g. $ make
MODULE=module-skeleton). The output files (namely the firmware binaries) are
written in a separate directory as well (output-MODULE).

This behavior enables the parallel development of multiple modules (or multiple
versions of the same module) within the same FDK instance.

The FDK build system has also been cleaned up, with obsolete rules removed and
comments added. The build.sh shell script has been moved to scripts/ so that
only the main Makefile is visible for the users.

The README.md was updated with new instructions.

Testing Done: compiled and tested with GPIO demo
Bug Number: SW-1708
Review URL: https://reviews.leaflabs.com/r/1992/
Reviewed by: fparent, mbolivar

Signed-off-by: Joel Porquet <joel@porquet.org>