Add support for VID/PID
Module developers need separate vendor/product_id for separate modules, but the
es2 hardware doesn't support that. These numbers are significant as
boot-over-unipro needs this to match firmware images from module store.

VID/PID values are to be given in the module.mk file of a module-specific
directory. The values are given to bootrom when generating the S1 bootloader
image, and to bootrom-tools when creating the tftf image of the firmware.

Testing Done: compiled
Bug Number: SW-1853
Review URL: https://reviews.leaflabs.com/r/2140/
Reviewed by: mbolivar, gregkh

Signed-off-by: Joel Porquet <joel@porquet.org>