build: change the output directory for modules
Instead of putting the generated objects in the source directory of a module and
the resulting images in a new directory "output-$(MODULE)" at the top dir of the
FDK, put everything under a common "build" directory instead.

This avoids having a mess of directories at the root of the FDK, and also having
generated objects in the same directory as the source code.

Testing Done: compiled
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2188/
Reviewed by: mbolivar, gregkh, esennesh

Signed-off-by: Joel Porquet <joel@porquet.org>