build: fix the copy of a module source code
In order to keep the module source code free of object generation, we copy it to
the build directory. For that we rely on $OOT_BOARD which contains the name of
the c source files to be considered. However we miss the possible header files.

This patch fixes this issue by bluntly copying the entire source code in order
to be sure we're not missing any files.

Testing Done: compiled
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2300/
Reviewed by: Mike Mogenson, gregkh

Signed-off-by: Joel Porquet <joel@porquet.org>