camera: Update to the latest CSI-RX API
The CSI-RX init function is meant to perform most of the initialization
of the CSI receiver, short of starting it. Call it when configuring the
streams as that is state transition where the module should perform as
much of the time consuming tasks as possible. Similarly call the uninit
function when unconfiguring the module.

The CSI-RX start and stop function are then called in the capture and
flush operation handlers to start and stop the video stream.

Testing Done: Tested on DB3 with the white camera module.
Bug Number: SW-1832
Review URL: https://reviews.leaflabs.com/r/2527/
Reviewed by: gregkh

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>