README.md: fix code blocks
Using quotes for code was not the best idea. It doesn't support multiline
blocks, and is not made for that anyway. Let's replace all the quoteblocks with
proper code blocks and have a more coherent look.

Testing Done: copied it in the github markdown editor and checked the output
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2612/
Reviewed by: Mark Greer, Mike Mogenson

Signed-off-by: Joel Porquet <joel@porquet.org>