The FDK nouveau has arrived!
Replace the difficult to maintain Makefile with a better, more flexible build
system script. This new FDK is able to generate images for frame firmware (SVC,
APBridgeA, APB2) and module firmware. It supports ES2 and ES3 outputs, and is
capable of generating S1 bootloader images.

Other features include:
- unified build log
- dry-run (show commands of the build but do not execute)
- keep intermediary files
- build all the firmware images at once (or only frame or module images)
- update all config files at once
- bonus: bash and zsh completion

Examples of usage:

$ ./fdk.sh build all-frame
... build all the frame firmware images ...

$ ./fdk.sh -n build module/skeleton/module.mk
... shows the commands for building the skeleton module ...

$ ./fdk.sh -h
... shows the help ...

Testing Done: compiled and tested on DB3.1 and DB3.5. We generated the images of
the frame components, pushed them on the board, rebooted, and were able to list
the modules with lsgb.
Bug Number: SW-2736
Review URL: https://reviews.leaflabs.com/r/2917/
Reviewed by: Marti Bolivar, Mark Greer

Signed-off-by: Joel Porquet <joel@porquet.org>