Update README.md
This file is not up to date as we're currently transitioning to a new FDK and to
an Android xml manifest (instead of using git submodules). Let's update it by
mentioning a link to the Confluence page that describes the workflow.

Testing Done: compiled
Bug Number: SW-2736
Review URL: https://reviews.leaflabs.com/r/2916/
Reviewed by: Marti Bolivar, Mark Greer

Signed-off-by: Joel Porquet <joel@porquet.org>