fdk.sh: separate tftf and ffff generation
Not knowing that ES3 required both tftf and ffff packages (and not only ffff),
the ffff creation was creating a temporary tftf file and crafting the ffff
filename based on the tftf file before removing the tftf file. Now that we want
to keep the tftf around, it's getting difficult to have several levels of
variable references.

With this patch, we perform the tftf generation first by explicitly calling the
tftf generation function, and then the ffff generation.

Testing Done: Compiled a few modules to check the generated ES2 and ES3 firmware
images
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2921/
Reviewed by: Jun Li, Eli Sennesh, Morgan Girling

Signed-off-by: Joel Porquet <porquet_joel@projectara.com>