scripts: add helper for balancing vid/pid values
This helper receives a value (or multiple values) on the command line and
transforms it into a hamming balanced value. If the value given on the command
line is already a 32-bit value, then it checks if the value is balanced or not.

This helper aims to help module developers to generate the balance vid/pid
values they have to put in their module.mk.

Testing Done: used for transforming the vid/pid from unbalanced to balanced in
the previous patch
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2923/
Reviewed by: Jun Li, Greg KH

Signed-off-by: Joel Porquet <porquet_joel@projectara.com>