fdk.sh: fix the get_var_mk() for the Android build
It seems that the 'make' command used for the Android build is not the standard
GNU make and produces some extra output that get in the way. Adding '--quiet'
solves the issue (and is compatible with the regular make so it's transparent
for everyone else).

Testing Done: compiled the skeleton module
Bug Number: SW-2952
Review URL: https://reviews.leaflabs.com/r/2984/
Reviewed by: Joel Porquet, Jeff Carlyle

Signed-off-by: Michael Scott <michael.scott@linaro.org>