fdk.sh: better error logging
Errors generated by commands should go into the build log as well, not only on
the console output. And developers should be made aware that a build.log is
accessible to get the reason of the failure (in case they didn't enable
verbosity).

Testing Done: compiled a module with a modification that generated an error and
tested that the error output ends up in the build.log file
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/3014/
Reviewed By: Marti Bolivar

Signed-off-by: Joel Porquet <porquet_joel@projectara.com>