fdk.sh: catch error when config is not defined in the mk target file
If 'config' is not defined in the mk target file, it's an issue and should be
raised early. Otherwise it breaks later when installing the file before
compiling nuttx and the error is difficult to understand.

Testing Done: compiled with config being undefined and checked that it failed
early with a relevant error message
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/3015/
Reviewed By: Marti Bolivar

Signed-off-by: Joel Porquet <porquet_joel@projectara.com>